Added VMethodInfoMap and VMethodInfo types to the RTTI module. Now there is a single instance for every vmethod (before we were just tracking the vmethod_ea in the VtableEntry)
Implemented basic save/load to/from a json file, for the RTTIDatabase. We could save this in a netnode, but I chose not to for simplicity and
not having to require an external dependency

collect vtables phase now attempts to fix function boundaries (like before). Still has more work to be done there.
Started finding the OSSafeMetaCast method