This commit implements applying the mangled symbols to the vmethods It includes work to associate a vmethod with its "owning" vtable entry. While working on this I revealed that some vmethods do not correspond to a single vtable entry (like I initially thought could be done).
Classes that do not have a vtable now marked as optimized and their
subclasse's vtable information construction had been fixed.

Added blacklists to the collect classes phase.Currently I do not handle
templated names well, and the existing code that does it is not good
enough.
The blacklist also contains classes that are some edge-cases that I wish to solve in the
future, and delay me from testing the current logic on most of the
classes in the kernelcache.
Seems like the only templated class in the kernelcache is the
OSValueObject (also present in the XNU sources..)

Improvements to the tests modules, added the `test_utils.py` which
provides an easy contextmanager to open a kernelcache and process it